plugins/moondream_vision/requirements.txt
# Moondream 视觉插件依赖（与仓库根目录 requirements.txt 分离）
# 安装：pip install -r plugins/moondream_vision/requirements.txt
#
# PyTorch：请按本机 CUDA 到 https://pytorch.org 选择命令；下列 torch 版本为下限示意。
transformers>=4.51.1
accelerate>=1.10.0
torch>=2.1.0
Pillow>=10.0.0
# INT8 / INT4（bitsandbytes）在 NVIDIA CUDA 环境下用于降低显存；未用量化可不装。
bitsandbytes>=0.43.1